Feature/new endpoints

+++ backend/src/resources/paymentMethod/paymentMethod.service.ts
@@ -1,14 +1,13 @@
-import { Request, Response } from "express";
-import { PaymentMethod, PrismaClient } from "@prisma/client";
+import { Request } from "express";
+import { PrismaClient } from "@prisma/client";
 import { PaymentMethodDto, CreatePaymentMethodDto } from "./paymentMethod.types";
 
 const prisma = new PrismaClient();
 
 export const createPaymentMethod = async (
   paymentMethod: CreatePaymentMethodDto,
-  req: Request,
+  uid: string,
 ): Promise<PaymentMethodDto> => {
-  const userId = req.session.uid as string;
   return await prisma.paymentMethod.create({
     select: {
       id: true,
@@ -19,17 +18,17 @@ export const createPaymentMethod = async (
     },
     data: {
       ...paymentMethod,
-      userId: userId,
+      userId: uid,
     },
   });
 };
 
 export const listPaymentMethods = async (
-  req: Request,
+  uid: string,
   skip?: number,
   take?: number,
 ): Promise<PaymentMethodDto[]> => {
-  const paymentMethods = await prisma.paymentMethod.findMany({
+  return prisma.paymentMethod.findMany({
     select: {
       id: true,
       type: true,
@@ -37,18 +36,17 @@ export const listPaymentMethods = async (
       createdAt: true,
       updatedAt: true,
     },
-    where: { userId: req.session.uid },
+    where: { userId: uid },
     skip,
     take,
   });
-  if (!paymentMethods) {
-    throw new Error("Nenhum método de pagamento foi encontrado");
-  }
-  return paymentMethods;
 };
 
-export const readPaymentMethod = async (id: string, req: Request): Promise<PaymentMethodDto> => {
-  const paymentMethod = await prisma.paymentMethod.findUnique({
+export const readPaymentMethod = async (
+  id: string,
+  uid: string,
+): Promise<PaymentMethodDto | null> => {
+  return await prisma.paymentMethod.findUnique({
     select: {
       id: true,
       type: true,
@@ -56,53 +54,21 @@ export const readPaymentMethod = async (id: string, req: Request): Promise<Payme
       createdAt: true,
       updatedAt: true,
     },
-    where: { id },
+    where: { id: id, userId: uid },
   });
-  if (!paymentMethod) {
-    throw new Error("Método de pagamento não encontrado");
-  }
-  if (paymentMethod.userId !== req.session.uid) {
-    throw new Error("Usuário não autorizado a visualizar este método de pagamento");
-  }
-  return paymentMethod;
 };
 
 export const updatePaymentMethod = async (
   id: string,
   updatedPaymentMethod: CreatePaymentMethodDto,
-  req: Request,
-): Promise<PaymentMethodDto> => {
-  const paymentMethod = await prisma.paymentMethod.findUnique({
-    select: { userId: true },
-    where: { id },
-  });
-
-  if (!paymentMethod) {
-    throw new Error("Método de pagamento não encontrado");
-  }
-
-  if (paymentMethod.userId !== req.session.uid) {
-    throw new Error("Usuário não autorizado a atualizar este método de pagamento");
-  }
-
+  uid: string,
+): Promise<PaymentMethodDto | null> => {
   return await prisma.paymentMethod.update({
-    where: { id },
+    where: { id: id, userId: uid },
     data: updatedPaymentMethod,
   });
 };
 
-export const deletePaymentMethod = async (id: string, req: Request): Promise<PaymentMethodDto> => {
-  const paymentMethod = await prisma.paymentMethod.findUnique({
-    select: { userId: true },
-    where: { id },
-  });
-
-  if (!paymentMethod) {
-    throw new Error("Método de pagamento não encontrado");
-  }
-
-  if (paymentMethod.userId !== req.session.uid) {
-    throw new Error("Usuário não autorizado a remover este método de pagamento");
-  }
-  return await prisma.paymentMethod.delete({ where: { id } });
+export const deletePaymentMethod = async (id: string, uid: string): Promise<PaymentMethodDto> => {
+  return await prisma.paymentMethod.delete({ where: { id: id, userId: uid } });
 };
